+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 2.15 (next release)
 ---------------------------
+* Made convexity checking recursive, instead of being cached.
 
 Version 2.14
 ------------